AEA-262: Refactored translation service into smaller modules.
