+++ Makefile
@@ -80,7 +80,6 @@ build-specification:
 build-coordinator:
 	npm run --prefix=coordinator/ build
 	cp coordinator/package.json coordinator/dist/
-	cp coordinator/src/resources/ConvertWrapper.mustache coordinator/dist/resources/
 	cp coordinator/src/resources/ebxml_request.mustache coordinator/dist/resources/
 	poetry run scripts/update_coordinator_tests.py
