+++ coordinator/src/routes/prescription/post-prepare-message.ts
@@ -1,6 +1,6 @@
 import {validatingHandler} from "../../services/handler";
-import * as translator from "../../services/translation-service";
-import {Bundle} from "../../services/fhir-resources";
+import * as translator from "../../services/translation/translation-service";
+import {Bundle} from "../../model/fhir-resources";
 import Hapi from "@hapi/hapi";
 
 export default [
@@ -13,7 +13,7 @@ export default [
         handler: validatingHandler(
             false,
             (requestPayload: Bundle, responseToolkit: Hapi.ResponseToolkit) => {
-                const response = translator.convertFhirMessageToHl7V3SignedInfo(requestPayload)
+                const response = translator.convertFhirMessageToHl7V3SignedInfoMessage(requestPayload)
                 return responseToolkit.response(response).code(200)
             }
         )
