+++ coordinator/tests/resources/test-resources.ts
@@ -1,8 +1,8 @@
 import * as XmlJs from 'xml-js'
 import * as fs from 'fs'
 import * as path from "path"
-import {ParentPrescription} from "../../src/services/hl7-v3-prescriptions";
-import {Bundle, Parameters} from "../../src/services/fhir-resources";
+import {ParentPrescription} from "../../src/model/hl7-v3-prescriptions";
+import {Bundle, Parameters} from "../../src/model/fhir-resources";
 import {ElementCompact} from "xml-js";
 
 export function clone<T>(input: T): T {
